Add files via upload
These are all of my python files that I included. I did not get them all, but that is where the points for the lab comes in. I feel that I should still receive credit for my portfolio here in GitHub because I followed the directions and did everything my repository needed.